+++ src/lib/krb5/os/dnsglue.c
@@ -184,8 +184,8 @@ krb5int_dns_nextans(struct krb5int_dns_state *ds,
         if (len < 0)
             return -1;
         ds->cur_ans++;
-        if (ds->nclass == ns_rr_class(rr)
-            && ds->ntype == ns_rr_type(rr)) {
+        if (ds->nclass == (int)ns_rr_class(rr)
+            && ds->ntype == (int)ns_rr_type(rr)) {
             *pp = ns_rr_rdata(rr);
             *lenp = ns_rr_rdlen(rr);
             return 0;
